# api-spec.yaml
openapi: 3.0.1
info:
  title: CirendsAPI
  version: "1.0"
servers:
  - url: https://cirends.runasp.net/api
  - url: http://localhost:5000/api
...
security:
  - Bearer: []
components:
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
      description: "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
...
paths:
...
  /activities/{activityId}/expenses:
    get:
      tags: [Expenses]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                - { id: 21, name: "Maistas", amount: 60.0, currency: "EUR", activityId: 5, paidByUserId: 1 }
        "404": { description: Not Found }
    post:
      tags: [Expenses]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            example:
              name: "Maistas"
              description: "…"
              amount: 60.00
              currency: "EUR"
              expenseDate: "2025-01-10T13:00:00Z"
              paidByUserId: 1
              shares:
                - { userId: 2, shareAmount: 30.00 }
                - { userId: 3, shareAmount: 30.00 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              example: { id: 21, name: "Maistas", amount: 60.0, currency: "EUR" }
        "400": { description: Bad Request }
        "404": { description: Not Found }
  /activities/{activityId}/expenses/{id}:
    get:
      tags: [Expenses]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              example: { id: 21, name: "Maistas", amount: 60.0, currency: "EUR" }
        "404": { description: Not Found }
    put:
      tags: [Expenses]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            example: { name: "Maistas (atnaujinta)", amount: 62.0 }
      responses:
        "204": { description: No Content }
        "400": { description: Bad Request }
        "404": { description: Not Found }
    delete:
      tags: [Expenses]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
        "404": { description: Not Found }
...